fix 有符号无符号混用

+++ services/cloudfiledaemon/src/fuse_manager/fuse_manager.cpp
@@ -1158,7 +1158,7 @@ static void CloudForgetMulti(fuse_req_t req, size_t count,
     fuse_reply_none(req);
 }
 
-static bool IsPageCached(shared_ptr<CloudInode> cInode, int64_t pageIndex)
+static bool IsPageCached(shared_ptr<CloudInode> cInode, uint64_t pageIndex)
 {
     if (!IsVideoType(cInode->mBase->name)) {
         return false;
@@ -1178,16 +1178,17 @@ static void HasCache(fuse_req_t req, fuse_ino_t ino, const void *inBuf)
     }
 
     const struct HmdfsHasCache *ioctlData = reinterpret_cast<const struct HmdfsHasCache *>(inBuf);
-    if (!ioctlData || ioctlData->offset < 0 || ioctlData->offset > cInode->mBase->size || ioctlData->readSize < 0) {
+    if (!ioctlData || ioctlData->offset < 0 || static_cast<uint64_t>(ioctlData->offset) > cInode->mBase->size ||
+        ioctlData->readSize < 0) {
         fuse_reply_err(req, EINVAL);
         CLOUD_FILE_FAULT_REPORT(CloudFileFaultInfo{PHOTOS_BUNDLE_NAME, FaultOperation::IOCTL,
             FaultType::WARNING, EINVAL, "invalid argument in ioctl"});
         return;
     }
-    int64_t headIndex = ioctlData->offset / MAX_READ_SIZE;
-    int64_t tailIndex = headIndex;
-    if (ioctlData->offset + MAX_READ_SIZE < cInode->mBase->size) {
-        tailIndex = (ioctlData->offset + MAX_READ_SIZE) / MAX_READ_SIZE;
+    uint64_t headIndex = static_cast<uint64_t>(ioctlData->offset) / MAX_READ_SIZE;
+    uint64_t tailIndex = headIndex;
+    if (static_cast<uint64_t>(ioctlData->offset) + MAX_READ_SIZE < cInode->mBase->size) {
+        tailIndex = (static_cast<uint64_t>(ioctlData->offset) + MAX_READ_SIZE) / MAX_READ_SIZE;
     } else {
         tailIndex = cInode->mBase->size / MAX_READ_SIZE;
     }
